requirements.txt
fastapi[all]
torch
torchvision
requests
Pillow
python-multipart
uvicorn
